Put Feedback events into a Concern
